+++ src/models/operations/createoutput.ts
@@ -11176,6 +11176,8 @@ export type OutputTcpjson = {
   textSecret?: string | undefined;
 };
 
+export type PqControlsWizHec = {};
+
 export type OutputWizHec = {
   /**
    * Unique ID for this output
@@ -11287,7 +11289,48 @@ export type OutputWizHec = {
   wiz_sourcetype: string;
   description?: string | undefined;
   /**
-   * Wiz Defender Auth token
+   * Use FIFO (first in, first out) processing. Disable to forward new events to receivers before queue is flushed.
+   */
+  pqStrictOrdering?: boolean | undefined;
+  /**
+   * Throttling rate (in events per second) to impose while writing to Destinations from PQ. Defaults to 0, which disables throttling.
+   */
+  pqRatePerSec?: number | undefined;
+  /**
+   * In Error mode, PQ writes events to the filesystem if the Destination is unavailable. In Backpressure mode, PQ writes events to the filesystem when it detects backpressure from the Destination. In Always On mode, PQ always writes events to the filesystem.
+   */
+  pqMode?: models.ModeOptions | undefined;
+  /**
+   * The maximum number of events to hold in memory before writing the events to disk
+   */
+  pqMaxBufferSize?: number | undefined;
+  /**
+   * How long (in seconds) to wait for backpressure to resolve before engaging the queue
+   */
+  pqMaxBackpressureSec?: number | undefined;
+  /**
+   * The maximum size to store in each queue file before closing and optionally compressing (KB, MB, etc.)
+   */
+  pqMaxFileSize?: string | undefined;
+  /**
+   * The maximum disk space that the queue can consume (as an average per Worker Process) before queueing stops. Enter a numeral with units of KB, MB, etc.
+   */
+  pqMaxSize?: string | undefined;
+  /**
+   * The location for the persistent queue files. To this field's value, the system will append: /<worker-id>/<output-id>.
+   */
+  pqPath?: string | undefined;
+  /**
+   * Codec to use to compress the persisted data
+   */
+  pqCompress?: models.CompressionOptionsPq | undefined;
+  /**
+   * How to handle events when the queue is exerting backpressure (full capacity or low disk). 'Block' is the same behavior as non-PQ blocking. 'Drop new data' throws away incoming data, while leaving the contents of the PQ unchanged.
+   */
+  pqOnBackpressure?: models.QueueFullBehaviorOptions | undefined;
+  pqControls?: PqControlsWizHec | undefined;
+  /**
+   * Wiz Defend Auth token
    */
   token?: string | undefined;
   /**
@@ -21133,6 +21176,24 @@ export function outputTcpjsonToJSON(outputTcpjson: OutputTcpjson): string {
 }
 
 /** @internal */
+export type PqControlsWizHec$Outbound = {};
+
+/** @internal */
+export const PqControlsWizHec$outboundSchema: z.ZodType<
+  PqControlsWizHec$Outbound,
+  z.ZodTypeDef,
+  PqControlsWizHec
+> = z.object({});
+
+export function pqControlsWizHecToJSON(
+  pqControlsWizHec: PqControlsWizHec,
+): string {
+  return JSON.stringify(
+    PqControlsWizHec$outboundSchema.parse(pqControlsWizHec),
+  );
+}
+
+/** @internal */
 export type OutputWizHec$Outbound = {
   id: string;
   type: "wiz_hec";
@@ -21169,6 +21230,17 @@ export type OutputWizHec$Outbound = {
   data_center: string;
   wiz_sourcetype: string;
   description?: string | undefined;
+  pqStrictOrdering?: boolean | undefined;
+  pqRatePerSec?: number | undefined;
+  pqMode?: string | undefined;
+  pqMaxBufferSize?: number | undefined;
+  pqMaxBackpressureSec?: number | undefined;
+  pqMaxFileSize?: string | undefined;
+  pqMaxSize?: string | undefined;
+  pqPath?: string | undefined;
+  pqCompress?: string | undefined;
+  pqOnBackpressure?: string | undefined;
+  pqControls?: PqControlsWizHec$Outbound | undefined;
   token?: string | undefined;
   textSecret?: string | undefined;
 };
@@ -21216,6 +21288,17 @@ export const OutputWizHec$outboundSchema: z.ZodType<
   data_center: z.string(),
   wiz_sourcetype: z.string(),
   description: z.string().optional(),
+  pqStrictOrdering: z.boolean().optional(),
+  pqRatePerSec: z.number().optional(),
+  pqMode: models.ModeOptions$outboundSchema.optional(),
+  pqMaxBufferSize: z.number().optional(),
+  pqMaxBackpressureSec: z.number().optional(),
+  pqMaxFileSize: z.string().optional(),
+  pqMaxSize: z.string().optional(),
+  pqPath: z.string().optional(),
+  pqCompress: models.CompressionOptionsPq$outboundSchema.optional(),
+  pqOnBackpressure: models.QueueFullBehaviorOptions$outboundSchema.optional(),
+  pqControls: z.lazy(() => PqControlsWizHec$outboundSchema).optional(),
   token: z.string().optional(),
   textSecret: z.string().optional(),
 });